+++ src/OpenLoco/src/S5/S5.cpp
@@ -419,7 +419,7 @@ namespace OpenLoco::S5
         dst = *exportGameState(src);
         dst.general.savedViewX = savedView.viewX;
         dst.general.savedViewY = savedView.viewY;
-        dst.general.savedViewZoom = static_cast<uint8_t>(savedView.zoomLevel);
+        dst.general.savedViewZoom = static_cast<uint8_t>(static_cast<int8_t>(savedView.zoomLevel));
         dst.general.savedViewRotation = savedView.rotation;
 
         // Copy tile elements; remove any ghosts before saving
